Implemented communication channels.
Communication channels have been implemented as ClusterShell events handlers.
Whenever a message chunk is read, the data is given to a SAX XML parser, that
will use it to create corresponding messages instances through a messages
factory.

As soon as an instance is ready, it is then passed to a "driver". The driver is
an interface able to get and emit messages instances. Subclassing this interface
offers the ability to connect whatever logic you want over a communication
channel.

Exit messages have been removed, are they become useless with the new <channel>
tag that encapsulate a whole session. Some comsmetic fixes have been done too.