Task: allow enabling/disabling MsgTree buffering at runtime
For now, it was only possible to disable task's internal output
gathering on stdout (or stderr) with task.set_default() when task was
not running. For example, before the first run:
    task.set_default("stdout_msgtree", False)
    task.run()

This patch allows programs to call task.set_default() at any time, to
disable or enable again output gathering. Note that by default,
gathering is enabled on stdout and also on stderr when it is separated
from stdout. Disabling internal output gathering when it's not useful
can save memory and CPU time.

Closes #181

Change-Id: I80d9d85eb34d24712b748a2115a77877b84a8930