Accecpt new options "ssh_path", "ssh_options", ... from commandline too. Closes #247.
.Test
> clush --ssh_path='sshpass -p <MySecretPassword> ssh' --ssh_options='-oBatchMode=no -oStrictHostKeyChecking=no' -l <MyUserLogin> -w <OtherHost> echo hallo

This possibility is important because users will use sshpass usualy once in the beginning of a session to append their certificates to the remote authorized_hosts file. Once this is done sshpass is not needed any more as clush can connect the usual way (which is much faster). In order to not be forced to modify the configuration file just for this one and first connect, users should be enabled to pass --ssh_path, --scp_path, --ssh_options and --scp_options on the command line.

.Use case: Distribute and install ssh-keys
> cat ~/.ssh/id_rsa.pub | clush --ssh_path='sshpass -p <MySecretPassword> ssh' --ssh_options='-oBatchMode=no -oStrictHostKeyChecking=no' -l <MyUser> -w <OtherHosts> -b 'cat - >>/home/getreu/.ssh/authorized_keys'