Engine: fix multiple delayed timers deadlock
Issue #254 has identified a use case of two or more long
running/blocking timers leading to a runloop deadlock. In that case, the
first timer fires and blocks, delaying other timers, which are always
expired as they are rearmed in the very near future and block again...

The problem was that we didn't process each timer once per runloop. We
now build a list of expired timers and process them carefully only once
per runloop, thus avoiding any rearming deadload or missed autoclose.

This patch also adds a new debug message in case of delayed rearmed
timers and simplify the way next fire_date is set in that case.

Fix #254.

Change-Id: I1b34d4dbec67957a2ae030dac93461cd9feb3a4a