Update patch set 3
Patch Set 3:

I was a bit lazy and added only the classes referenced by the docs and the guide in particular. For example, we reference some NodeSetBase methods that are not overrided in NodeSet, sphinx doesn't seem to resolve inheritance. But overall thanks for this review, I agree with everything and it's very useful. I will work on it ASAP.

Patch-set: 3
Label: Code-Review=0