Update patch set 2
Patch Set 2:

(1 comment)

Aurélien Degrémont remarked that this patch is not as efficient as the previous way of trimming whitespaces... Do you want to keep working with this method of trimming once and for all the whitespaces? Or should I just work on a more efficient one?

Patch-set: 2
Label: Code-Review=0