Update patch set 4
Patch Set 4:

When you first submitted this patch, I checked the best way to do what you want, but it is not easy at all. In theory, splitting between operators being already done by _next_op(), this should be the place to strip, but it's a mess to do it. Finally, I accepted the idea of pre-scanning the string with split/joining. But if you noticed a performance regression, that's very annoying...
Also, I didn't have time to look in details your other patches, but I don't like much this additional strip argument.

Patch-set: 4
Label: Code-Review=0