Update patch set 7
Patch Set 7:

(2 comments)

Need to patch vim syntax file for to display coma with a different color. See (https://github.com/cea-hpc/milkcheck/blob/master/doc/vim/syntax/milkcheck.vim) as an example.

Files in 'groupfiles/' directory will have 2 kind of syntax: with or without headers: "[sourcename]"
I have in mind ideas to do automatic parsing of 'groupfiles/' content. It could be nice if we have a clear statement on this directory content.
I think we are on the edge of doing something more generic and powerful and I would like to avoid introduce incompatibilities we will have to support in the future.

Patch-set: 7