Update patch set 2
Patch Set 2: Code-Review+1

Looks fine, in spite of the dictionaries roughly exposed to the external world. I don't find the interface very handy and would suggest a get/set API handling key, value and scope.

I know that it is recommended to avoid getters/setters on simple attributes. But dictionaries lay at the limit between "simple" and "not-so-simple" attributes...

What do you think? Would it be overkill?

Patch-set: 2
Label: Code-Review=+1