Update patch set 2
Patch Set 2:

(1 comment)

I like the global idea. Maybe we just need to make this private. Or just undocumented for now with no guarantee the API will be stable.

Simply do not document this in the guide, and add in pydoc for this class that this internal only, for now. Task.set_info()/set_default() are still usable anyway.

Patch-set: 2