Tree: support multihomed gateways
If the Gateway's system hostname cannot be found in received
TopologyTree, the Gateway will iterate over its local interfaces using
the netifaces module and get their associated hostnames to find an
alternate hostname in the tree. The matching hostname will then be used
as new root of the PropagationTreeRouter instead of system hostname.

To enable this feature, the netifaces Python module is required on the
gateway (version 0.5 or above: https://pypi.python.org/pypi/netifaces ).

Closes #250.

Change-Id: Ide88032e99e44071b2da33a6e507a521ee029b30