Tree: support multihomed gateways
Fixed a tree issue when hosts provided in topology.conf don't match the
gateway system hostnames. The gateway host/alias found in topology.conf
is now sent during channel configuration. It is then received by the
gateway and used as new root of the PropagationTreeRouter instead of
system hostname. Both parent node and gateway should be upgraded to
benefit from this feature, however, not doing so shouldn't break
existing installation.

Closes #250.

Change-Id: Ide88032e99e44071b2da33a6e507a521ee029b30