Update patch set 11
Patch Set 11: Code-Review+1

Needs Verified - I GOT IT!!! Thanks Dominique for the idea. It was even more complicated due to some tricky cases. And there is a mistake in your example, but not in nodeset anymore:

$ nodeset --autostep=10% -f n1[1-9,15,59,10-50,142]0
n[110-190/10,1100-1500/10,1590,11420]


Also I found out a very tricky case is:

$ nodeset -f node-1[0-48/8]
node-[10,18,116,124,132,140,148]

versus the 0-padding version:

$ nodeset -f node-1[00-48/8]
node-[100,108,116,124,132,140,148]

But it works, yeess :)
Please note that while the code is quite big, it is only triggered when nodesets are misformated, so it shouldn't impact most of existing app performance (I hope so).



Padding mismatch is my other concern now, example:

$ nodeset -f prod-02[10-34,069-099]
prod-[02010-02034,02069-02099]

it is not supported, but with the new code, it leads to erroneous indexes, note the 02010 -> this is due to the pre-padding of the inner range before amending... I think we can consider this case as illegal, maybe even we should raise an error, but we never did in the past and that could break (maybe?) existing code...
example:
$ nodeset -f prod-[01-2]
prod-[01-02]

but that may be part of another ticket!

Patch-set: 11