Update patch set 6
Patch Set 6:

The i stands for inet/inet6, not iPad :D (I know, I can do better)

Thanks for checking. Indeed, simple CNAME/aliases won't suffice but that's in purpose. I think the idea behind hostname checking is a sanity check when propagating command. Maybe Henri can comment on that too, there might be other reasons?  inet hosts checking is another way to ensure that the gateway is really the right command router.

I still think that this feature should allow everyone having this multiple interfaces issue to fix it, eg. just by adding an entry to /etc/hosts for the local interface. That is was I used for testing and is very easy to use. As for the multiple DNS, indeed there might be a problem, but I'm almost sure /etc/hosts ( + file dns in nsswitch.conf ) can still help.

Still, if that definitely won't work, I don't see any reason we couldn't send a "fake" hostname in ConfigurationMessage (ie. the one we used to ssh-connect to the gateway == the one in the topology.conf). If we go this way, we lost the hostname checking feature for the newroot resolution. I'm not sure I like that much. Sanity checks are good :) I will think about it today.

Patch-set: 6