Update patch set 6
Patch Set 6:

Regular users cannot modify /etc/hosts.

I would like them to be able to use this feature whatever the system is. Right now, ClusterShell has no constraint on DNS or whatever the topology is. 

If this is only a sanity check, this is bad check :) as it breaks correct configuration.

From my point view, this detection is only needed to know, where the gateway is in the full topology it receives from the upper node. If we can easily tell the gateway, which gateway it is in the tree, let's do it and all problem will be solved :)

Patch-set: 6