Update patch set 7
Patch Set 7:

(3 comments)

Strangely, you implemented a protocol versioning system but you are not using it in your patch...

By example, Shine uses an autodetect mechanism for minor protocol changes, more or less introspection of messages. Protocol compat is managed in the only unpacking method, and in others unpickle methods.
Shine is at protocol version 3 for a while now. This will change only if the protocol is very different.

You can do as you like, but maybe avoid introducing this feature and not using it :)

Patch-set: 7
Label: Code-Review=0