Update patch set 3
Patch Set 3:

I was not thinking about putting this in documentation.
This is volatile, transient info :)
I was thinking of preparing this info for mailing list announce, etc.

By the way, there is no reference to Lib improvements in this text, neither "and a lot of bugfixes..." :)

Patch-set: 3
Label: Code-Review=0