Update patch set 1
Patch Set 1:

(6 comments)

A ticket is probably a better to discuss this feature and implementation details.

If I understand correctly, you need a way to select a subset of a nodeset, and do this, at random, or you want to "shuffle" the nodeset?

Keep in mind that NodeSet are always sorted, by design. So this is more an output option in my opinion.

Patch-set: 1
Label: Code-Review=0