Update patch set 2
Patch Set 2:

Not sure this is a good idea.
That means a lot of test IDs will have to be updated each time we had a new test to the main class. This is very painful to maintain, and rather useless.

Also, this breaks test history in tools like Jenkins.

Either:
- starts numbering at "100" for subclasses (not very nice)
- drop the numbering

Patch-set: 2