WIP: Engine: implement per worker fanout
This patch adds as Worker.fanout attribute. If defined, this value is
used instead of the engine fanout, and the worker's clients won't be
taken into account by the engine. The engine has its own counter for
clients with workers not using worker.fanout (default).

Additionally, the special value -1 set the worker fanout value to
unlimited, this makes it suitable for gateway channels and also gateway
stdin StreamWorker.

The worker.fanout value may be changed at any time (from the current
task), and 0 could be used, or None could be set back to use the
engine's fanout again, but some limitations may apply.

Fixes #322.

Change-Id: I343b270a812f5f0a9485e837aa608b1426ab422b