WIP: Engine: implement per worker fanout
This patch adds as Worker.fanout attribute. If defined, this value is
used instead of the engine fanout. The engine now keeps per-worker
registered client counters up-to-date. The engine also maintains a
default counter for workers that don't define worker.fanout (None).

Additionally, the special value FANOUT_UNLIMITED sets the worker fanout
value to unlimited, this makes it suitable for gateway channels and also
gateway stdin StreamWorker.

The current implementation doesn't allow to modify worker.fanout when
the worker has already been scheduled and the task is running.

Fixes #322.

Change-Id: I343b270a812f5f0a9485e837aa608b1426ab422b
Signed-off-by: Stephane Thiell <sthiell@stanford.edu>
Signed-off-by: Aurelien Degremont <aurelien.degremont@cea.fr>