Update patch set 2
Patch Set 2:

(3 comments)

That's really good. I know this is changing tests but very slightly. This is very safe. there is couple of code I would like to get rid of and this was an occasion to kick a part of it :)
This is very safe for tests.

Patch-set: 2
Label: Code-Review=0