MsgTree: Python 3 support (#337)
In Python 3, replace imap() with map() as it returns an iterator.

Use list comprehensions instead of filter() and map() in walk*()
methods, as the function parameter of both filter() and map() doesnt
accept a value of None anymore.

Use filterfalse instead of ifilterfalse whenever possible.

Python 3 support for dict iterators with possible performance overhead
in Python 2.

Change-Id: Id9dd66e6323655796da40a931e054b948f63a816