Task: add stdin boolean argument to shell() and run()
Add this convenience option to prevent the use of stdin. If set to
False, EOF is sent at first read. This is similar to what ssh -n does,
but because clustershell is able to easily send EOF itself at first
read, we can actually avoid connecting stdin to /dev/null.

The default value of this option is managed by Defaults. The default
value of the Defaults' value is set to True for compatibility.

Finally, this patch modifies clush to change the default value of stdin
if --nostdin is specified. The following case shouldn't block anymore:

    $ cat | clush --nostdin -w node cat

Closes #333.

Change-Id: I1560f367962c79713e76e07def86c67a39928e90