Update patch set 2
Patch Set 2:

This looks quite good. I'm not so familiar with bytes/string issue with Py3 so I cannot be 100% sure.

Does this mean MsgTree will be a bytes tree and never a string tree? In another world, where Python compat does not exist, is it possible to have a str MsgTree, using unicode, or this does not make sense at allt?

Are command (popen) output bytes or string?

Any additional explanation for dummies in commit message is appreciated.

Patch-set: 2