Update patch set 2
Patch Set 2:

ok in brief, it's simple. bytes is like str in Python 2, it's like a char*, no encoding attached. It's perfect for use as buffers and this is what we want for mostly everything in clustershell.  The cool thing is that bytes type does exist in Python 2 as an alias of the old str, so a common code base is possible. When you use Popen with Python 3, you get buffers in bytes instead of (encoded) string indeed. Python3's str are strings with encoding attached (a bit like the unicode type in Python 2). The only case where we need to deal with text encoding is when s is a str and hardcoded in clustershell, so we need  to use s.encode('utf-8') . Like for example in Display.py, there are several hardcoded strings like the header, in that case we'll need to convert that to bytes using a known encoding before displaying. Because print() can only print strings (with an encoding), we'll use sys.stdout.buffer in Python 3 which is similar to using sys.stdout in Python 2, and there is no encoding involved.

Patch-set: 2