EngineTimer: Python 3 support (#337)
An interesting change here... we have been using this kind of comparison
for a while now:

    None > -1.0e-3 (-EPSILON)

which returns False in Python 2. Timers initialized with None weren't
armed. This is what we wanted but can be super confusing and is not
allowed in Python 3 anymore.

This patch cleans things up and also implement the required comparison
operators to replace __cmp__() which is not supported/used in Python 3.

Most of the tests in TaskTimerTest.py now pass, the failed ones are not
related to timers.

Change-Id: I4fe65505acff2d1c6f62609a276a82d137044f2e