Update patch set 1
Patch Set 1:

(9 comments)

Hey! That was a big review! I spent a lot of time to understand this code to be able to do this review.

I ended up doing an audit of this timer management classes! ^^
At the end, I'm not sure these 2 new methods would really help make the code clearer. But there is plenty of room for code refactoring!
And THAT is what is needed to clarify and clean this code!

Tell me if I missed something big!

Patch-set: 1
Reviewer: Aurelien Degremont <1004434@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0