NodeUtils: use string as key for resolver caching (#337)
NodeSet is mutable, thus not hashable, but it looks like Python 2 wasn't
very strict about it.

Python 3 complains with the following exception:
  ...
  File "../lib/ClusterShell/NodeUtils.py", line 260, in resolv_reverse
    NODE=node)
  File "../lib/ClusterShell/NodeUtils.py", line 219, in _upcall_cache
    if key in cache and cache[key][1] < time.time():
nose.proxy.TypeError: unhashable type: 'NodeSetBase'

This patch adds an explicit cast to string for the key, which does
nothing unless a NodeSet or NodeSetBase are used as the key instead.
For instance, this does occur when resolv_reverse() is called.

Change-Id: If4c5c1083b79a0f96ca4aa8e8a7763d1170a8b14