Update patch set 1
Patch Set 1:

(1 comment)

I've been working on a couple of CS scripts were groups performances were the bottleneck. So I'm trying to get this code path as fast as possible.

We know NodeSet conversion from or to str is expansive. So I'm racking my brains to have the best way to handle it...

Patch-set: 1