NodeUtils: fix external reverse node group upcall caching issue
This patch adds an explicit cast to string for the cache key in
UpcallGroupSource.resolv_reverse() and adds a new test to cover
external reverse caching.

Indeed NodeSet was used as caching dict key, even though it is not
hashable. Note that external reverse caching is not used very often
(must be defined and the number of groups must be greater than the
number of nodes).

Change-Id: If4c5c1083b79a0f96ca4aa8e8a7763d1170a8b14