Worker: fix issue with retcode set to None
Do not propagate rc=None to the Task. None is not a valid retcode. rc
must be properly defined at the task level. The StreamClient uses rc =
None when closing, to trigger ev_hup etc. but there is no need to update
the task.

Somehow revealed by the Python 3 porting effort in tree mode. Triggers a
TypeError exception in Python 3 only:

  File "lib/ClusterShell/Task.py", line 1039, in _rc_set
      if self._max_rc is None or rc > self._max_rc:
      TypeError: unorderable types: NoneType() > int()

Indeed the issue is not seen in Python 2 because the comparison between
None and integers is supported.

Added specific test.

Change-Id: Ie165b5221e91efafb245c2f1ee7a071972e28ad4