Worker.Tree: fix selection of remote Python executable (#337)
The Python interpeter on gateways must be of the same major version than
on the root node for the tree mode to work. Use the same executable name
(eg. python, python2, python3) by default.

This is due to changes made to the pickle protocol:

https://docs.python.org/3/library/pickle.html#data-stream-format

Also slashes (/) are not supported in module name (-m), so use dot (.)
instead.

Finally, check the following environment variable:

    CLUSTERSHELL_GW_PYTHON_EXECUTABLE

that can be used to override the default selection, giving users an
option to provide an absolute path to the remote python executable.

Change-Id: If8af8abb202ec13fb7db063859f65283b4923d64