tree: NodeSet is not hashable, do not use it as dict key!
NodeSet is not hashable, and was used in numerous places as a dict key.
This is wrong and even fails in Python 3. This patch fixes this.

In PropagationTreeRouter, self.table is now a tuple of (key, value)
pairs instead of a dict.

For other cases, we use the string representation of the NodeSet as the
dict key.

Change-Id: Ie48e3993551475dcc424dff1d6fee4c67591cc1f