NodeUtils: support for lazy init in GroupResolverConfig
Add support for lazy initialization of the full node group configuration
in both GroupResolver and GroupResolverConfig.

This is especially useful when there are many group sources and yaml
files. Without this patch, NodeSet always loads the full group
configuration, even when not using node groups at all.

Derivative work of #348.

Change-Id: I27fd5d899b89b64b54fb0cd0aa3e0a8a19139965