Packaging improvements (added manifest template)
Python packaging:

Added a manifest template "MANIFEST.in" to allow the use of python
setup.py sdist in mkrpm.sh script to build the base source tarball.
The direct result of using sdist is the use of CamelCase in the base
source tarball's name (it matches the ClusterShell module name). This
change also allows the use of "python setup.py sdist upload" to update
PyPI from command line.

RPM packaging:

Fixed RPM specfile to take this new source tarball into account and use
PyPI as Source URL. Also misc fixes in the specfile to conform to the
Fedora packaging guidelines:
- removed BuildRoot tag
- removed Group tag
- do not clean %{buildroot} in %install

Change-Id: I6f6832c796fdf94fa3ec75c62914f1f59cc691b2