Packaging: major refresh for 1.8
Created separate Python 2 and 3 packages:

Use separate packages for Python 2 and Python 3 ClusterShell modules
and related command line tools. The python 2 and python 3 stacks are
intended to be fully-installable in parallel as recommended by the
Fedora/EPEL guidelines.

The "clustershell" package includes the common configuration files and
documentation and requires python2-clustershell, because Python 2 is the
default interpreter in EPEL 7.

Indeed, the provided specfile template does primarily target EPEL 7 but
can be easily modified to set python3-clustershell as the default, as
recommended in latest Fedora versions IF the command line tools do
actually provide the SAME functionality, which is _almost_ the case here
(tree communication protocols are not compatible).

Removed vim-clustershell subpackage:

vim-clustershell was confusing so we remove it and add the vim
extensions to the main clustershell subpackage. Another change is the
use of vim-filesystem instead of vim-common as a requirement. This is
needed for proper vim extension directory ownership.

Improved portability:

Tested with EPEL 6, 7, Fedora 26 and openSUSE Leap 42.3

Change-Id: I88411e5f2518e86acd44686d14aac2e14876c08a