Packaging: create separate Python 2 and 3 packages
Create separate packages for Python 2 and Python 3 ClusterShell modules
and related command line tools. The python 2 and python 3 stacks are
intended to be fully-installable in parallel as recommended by the
Fedora/EPEL guidelines.

The "clustershell" package includes the common configuration files and
documentation and requires python2-clustershell, because Python 2 is the
default interpreter in EPEL 7.

Indeed, the provided specfile template does primarily target EPEL 7 but
can be easily modified to set python3-clustershell as the default, as
recommended in latest Fedora versions IF the command line tools do
actually provide the SAME functionality, which is _almost_ the case here
(tree communication protocols are not compatible).

Change-Id: I88411e5f2518e86acd44686d14aac2e14876c08a