Update patch set 1
Patch Set 1:

Oh no!!! Python 3 FTW!!! And it's common practice as I have noticed. This specfile almost works on SLES. I will submit an updated patch to make this specfile works on el6, el7, Fedora and Leap, it's just a matter of missing macros. For me it is less work if I can use a single base specfile for everything.

Patch-set: 1