Update patch set 5
Patch Set 5:

That makes sense! PEP-4 seems to say the same: the old stuff should be documented as obsolete, and/then add a deprecation warning. But I would say DeprecationWarnings shouldn't be so noisy but they still are in Python 2.6 by default, unfortunately. I worked quite a long time on this patch, so I will keep it for 1.9. :) I'll rework the patch so that we can still have the new API tests.

Patch-set: 5
Label: Code-Review=0