Update patch set 13
Patch Set 13:

This code is kind of "temporary", so this is not such a big deal if not perfect IMO. I do not want to "hide" this kind of compat code behind a wrapper. Or this wrapper should have 'compat' in its name :)

You can us such wrapper is you want, bu it won't work for msgline use cases though. 
Calls are always done using positional args. So *args should be enough.

Patch-set: 13
Label: Code-Review=0