NodeUtils: ignore YAML group files with permission error (#348)
Sometimes users don't have access to some YAML group files. The current
behavior is a fatal error in that case, making clush or nodeset unusable
in some cases. For example, at the SRCC, we have operators without root
access that use clush and cluset on the cluster to gather cluster
accounting and other info, but they don't have access to all group files.
This change just ignores the YAML files the user doesn't have access to,
print a debug message and continue processing the rest.

Closes #348.

Change-Id: I0e8a82304ad592a2f0866b542191ebe1912e642d