Update patch set 1
Patch Set 1: Code-Review+2

Re: wildcard import, for TLib, I agree this could be removed. But for the rest, I feel like it should be tested somehow, somewhere. Python allows wildcard imports and even if it is a bad practice, users can use that to import (__all__) symbols. I don't know what are the best practices to test for non-regression in that case (say we add a variable that create a conflict for example?). But it can probably be removed from most of the tests indeed.

Patch-set: 1
Label: Code-Review=+2