Update patch set 1
Patch Set 1:

Especially if we don't define __all__ , shouldn't we test wildcard import, because maybe we could add more __all__ to avoid an existing symbol conflict? This is not really testing the language in that case but more testing what symbols we're "exporting". This is not very important but as wildcard import is possible, shouldn't we 'support' it ( 'test' it)?

Patch-set: 1