NodeUtils: check that yaml dict key type is string (#361)
YAML syntax is more powerful than other configuration files, which can
be confusing for users not used to its syntax. Switching to PyYAML
BaseLoader use strings but breaks other things so it doesn't seem to be
an option here. This patch adds a check to ensure that the type of group
source names and group names is string (eg. properly quoted if there are
only digits) and raise an GroupResolverConfigError with a proper error
message if not.

Closes #361

Change-Id: If52b726b29490c5f5f7ef97a2b9256280a3055a0