Update patch set 1
Patch Set 1:

I did ;)

More seriously, I need to check, that sounds tough. We already have unpickling compat tests for some cases but in that case it has to go through the WorkerTree code, which is currently not tested at all. That would be a great thing to have in the future..

Patch-set: 1
Label: Code-Review=0