NodeSet: implement node wildcard expansion (#349)
Node wildcard support, using the '*' and '?' special characters, allows
users to filter on node names that match the provided wildcard mask.

The implementation searches for nodes from 'all' nodes, that is, the
result of NodeSet.fromall(), nodes that are returned by the all upcall,
or in case of absence, the union of list+map.

With this change,

    $ nodeset -f '*'

becomes the same as:

    $ nodeset -f -a

Note that mixing wildcards and ranges (like '*foo[1-3]') is not
supported.

Closes #349

Change-Id: I1b99811dc65bd7fd9a5a9fedc6e5f1f3b8cb875e