NodeSet: implement node wildcard expansion (#349)
Node wildcard support, using the '*' and '?' special characters, allows
user to filter by node names that match the provided pattern.

The implementation searches for nodes from 'all' nodes, that is, the
result of NodeSet.fromall(), nodes that are returned by the all upcall,
or in case of absence, the union of list+map.

With this change,

    $ nodeset -f '*'

becomes the same as:

    $ nodeset -f -a

Closes #349

Change-Id: I1b99811dc65bd7fd9a5a9fedc6e5f1f3b8cb875e