Tree: support offline gateways (#260)
With this change, WorkerTree is now able to safely repropagate commands when
the communication channel to a specific gateway cannot be established at all.
Such offline gateways were already marked as unreachable, but the 'relaunch'
code was missing. A new WorkerTree._relaunch() method has been added for that
purpose.

Note that if a gateway fails once a propagation channel is "set up", the remote
commands cannot be safely repropagated at this point and so will be seen as
failed, but this was already the case.

Closes #260

Change-Id: I2be52b01638de95f58cedc0f2671f02cfffe8297