Update patch set 2
Patch Set 2: Code-Review+1

Not an expert of this part of the code.
So I'm trusting you. My general feeling is that there is too many cross references between classes (client <-> engine) and this should be cleaned in the future, but this is for another patch :)

Patch-set: 2
Reviewer: Aurelien Degremont <1004434@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1