Update patch set 1
Patch Set 1:

Looks ok, but it would be nice to fix #384 differently just before this release, in order to being able to install 1.8.1 on Windows with pip.

I think we just need to test for Unix before calling os.geteuid(), or just use a try catch and ignore the error.
What do you think?

Patch-set: 1
CC: Aurelien Degremont <1004434@d5d70762-12d0-45a1-890d-524b12d3f735>