Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Looks ok, but it would be nice to fix #384 differently just before this release, in order to being able to install 1.8.1 on Windows with pip.
> 
> I think we just need to test for Unix before calling os.geteuid(), or just use a try catch and ignore the error.
> What do you think?

I have been discussing this internally to get feedback and now we don't think this is a good idea to just ignore some Python modules on Windows, as it may be confusing for the users. And silently ignoring modules for Windows will be confusing... We agreed that a better solution is to be explicit and put the portable NodeSet part in a separate subpackage, that could be installed either on Windows or Linux. Do you really think it's overkill?

Patch-set: 1
Reviewer: Stephane Thiell <1003524@d5d70762-12d0-45a1-890d-524b12d3f735>