Defaults: add NodeSet's default folding axis (#401)
Add 'fold_axis' as a tunable of Defaults. If set, this is used by
NodeSet to disengage some nD folding. By default, nD folding is
tried on all axes.

A global Defaults is instantiated by the library, whose values can be
set in a config file. Below is an example of use to improve
compatiblity with Slurm:

/etc/clustershell/defaults.conf:

    [nodeset]
    fold_axis: -1

Once set, by default, the library will only fold node sets using the
last axis found.

Closes #401.

Change-Id: I41637c8e18567ae254dfb2118fe2e4420789f802