EngineClient: EnginePort improvements
This patch introduces several changes related to EnginePort.

Remove dependency with Task: indeed, as an EngineClient, an EnginePort
should not have any dependency with Task, so it can be used internally
by workers before a Task is even defined.

Generate ev_start event when an EnginePort is starting, so that we can
be notified with registered EnginePort when the Engine is starting.

Move 'port_qlimit' config default to an [engine] section in
defaults.conf so that it doesn't depend on Task anymore.