tests: fix StreamWorker test leading to bad Engine state
An unconfigured StreamWorker does not have streams to register in the
engine, so when registered, it stays until it is explicitly aborted. To
make that worse, it is using a 'fanout slot' and if the Engine's fanout
is too low, it can lead to the engine run loop exiting prematurely. This
is not a case handled very well by the Engine. See #497 for more details
about the problem.

In the meantime, we fix the test to exit with a clean engine state.

Fixes #488.