RangeSet: support negative ranges (#515)
Support negative ranges in RangeSet, with as little impact as
possible to the normal (positive range) codepath. Zero padding is not
available with negative ranges at this time for this reason.

In case of negative ranges, RangeSet will case the numbers to integer.
Internal sorting then looks like this (eg. -5-6):
['-5', '-4', '-3', '-2', '-1', '0', '1', '2', '3', '4', '5', '6']

Apart from the internal sorting change, little modifications are needed
in the parser and the folding method, so I guess it's worth it.

Usage examples:

    >>> RangeSet(range(-5,7)) - RangeSet(range(1,3))
    -5-0,3-6

    >>> list((RangeSet(range(-5,7)) - RangeSet(range(1,3))).intiter())
    [-5, -4, -3, -2, -1, 0, 3, 4, 5, 6]

    >>> list((RangeSet(range(-5,7)) - RangeSet(range(1,3))).contiguous())
    [-5-0, 3-6]

Hence, the following is now possible:

$ cluset -f foo[-10-10]
foo[-10-10]

$ cluset -f foo[-100--90/2]
foo[-100,-98,-96,-94,-92,-90]

Closes #515.